feat：Move `WeakRefVirtualProxy` to separate file
